avoid race conditions in storage
Our app has faced 58 crashes since we started using Skylab. I see that the access to the map property isn't synchronised.

For example, DefaultSkylabClient.fetchAll works with map on a background thread. At the same time the client may access map from another thread. It leads to a race condition, which mean lead to a crash

# Conflicts:
#	Sources/Skylab/UserDefaultsStorage.swift